Added org-journal package. Moved org-docs var to support.
Added org-journal package config to configure-project.el.

Moved spydez/dir/org-docs from finalize step up to dev-directories.el to be
ready when org-journal wants to access it and also be device dependant.

Added some notes to configure-shell from attempt number next to get it working
on windows with MinGW... Not great success but it's not basically totally broken
anymore.