Signatures,
Added configure-signatures.el.
Added a general `spydez/group' group for defcustoms.
Added a half dozen different signatures for different uses.
Added `spydez/signature/insert' and `spydez/signature/search' functions for
making and looking for signatures.

Bumped `whitespace-line' face up to spydez/theme/color/zenburn-magenta-bg-1 for
a tiny bit more pop from the background. I was second guessing some
'one-character-overs' some of the time.

Added a link to a nice unicode table website in init.el.

Moved old 'org-journal' hydra to be an 'org-mode' hydra, which is already a
misnomer as it's an 'org-journal and signatures' hydra. But sigs are only really
intended for testing out in org files right now so eh.