Added ability to piggyback on require. Fixed taskspace bug.
(spydez/info/require 'foo) will now also look for 'foo-secret and load it after
'foo. This is so defaults or public stuff can be provided in .emacs.d, and then
tweaked, overridden, or other stuff added as needed by the piggybacker. As
opposed to having to override the whole feature.

Fixed bug in taskspace DWIMs where input arg was incorrectly parsed.