Fixed spydez/magit/auto-commit to load magit if needed.
`spydez/magit/auto-commit' failed due to not-loaded magit functions if magit's
auto-loads hadn't been hit yet. Added a require to function to fix that.

Renamed `spydez/to-unix-auto' to let `spydez/magit/auto-commit' be the only
'auto' right now. Also it was an ancient naming scheme, so it and it's friend
were updated to "spydez/buffer/to-unix(-utf8)".

`spydez/delete-word' and `spydez/backward-delete-word' docstrings updated to
include most important feature - no kill ring interaction.