Added smarter beginning and end of visual line funcs.
`spydez/smarter-beginning-of-visual-line' replaces `beginning-of-visual-line'
with functionality similar to `spydez/smarter-move-beginning-of-line'. Point
will now move to start of visual line, then call
`spydez/smarter-move-beginning-of-line' for more options if that did nothing.

So "C-a" in e.g. org-mode will now move to:
  1) beginning of visual line
  2) beginning of logical line
  3) indentation of logical line

Similarly for "C-e", new `spydez/smarter-end-of-visual-line' replaces
`end-of-visual-line'. This will move point to:
  1) end of visual line
  2) end of logical line