Deadgrep fix for unicode on Windows (7).
Fixed my init to let deadgrep work finally when searching for unicode glyphs.
E.G. before 'M-x deadgrep RET §' would fail with:

  /usr/bin/bash: rg --color=ansi --line-number --no-heading --with-filename
  --fixed-strings --smart-case   -- \Â§ .: No such file or directory

  exited abnormally with code 127

This was a combination of me wanting to use bash in emacs, emacs expecting cmd
because windows, and emacs fucking up half of what it expected because UTF-8
settings, I think?

So for now try not-utf8 and cmd as default shell for things and... see what else
this breaks.

But at least I can search for my unicode marker now.