Added bufler package for buffer management. Disabled yas-snippets.
Added bufler for buffer management. Added it as a source to helm-mini as well.

Disabled yasnippet-snippets as they mostly just get in the way by
auto-completing a letter or two into a template I didn't want instead of the
variable name I did want.

Added some themes to try out into working-on.org.