mis2 work - messages, data, unit tests
Renamed 'parts' to 'content' for better indication of what it is: the contents
of the message, not the settings, style, or final actual message.

Created mis2//message/output* funcitions for piping final message to final
sinks.

Created a very stupid and simple mis2/message, for now. Will build on as time
goes on and features are added.

Renamed 'internal' to 'private' in some places, and to 'data' in others,
depending on if it was related to internal/private helpers, getters, etc or if
it was related to mis2's internal data format for passing between functions.

Offensively stupid mis2//build functions for now so I can get a deadly-stupid
mis2//message test in place as a starting point. Will need to break out into its
own file and everything soon.

Getters for mis2 data, settings.

Mocked out `mis2//message/output/to-*' for unit tests so we can intercept
message and check.

Commented out some not-yet-used code. Will delete or use eventually.

Added test for mis2//data/get, mis2//settings/get/from-data.