Converting to more flexable dir/file paths.
This step: making the more flexable solution and replacing all the old
constants.

Removed a bunch of `spydez/dir/___' variables.

Added `spydez/dirky/___' functions/vars to zeroth-directories.el.
Keywords for specific folders get grouped by a 'keybase'.
Can then ask for a path like so: (spydez/dirky/path :emacs :docs)

If it's a domain-specific path, you can leave the 'keybase' nil for it to
decided which domain you're in: (spydez/dirky/path nil :logbook)

master-list.el sets this 'current domain' keyword default when it figures out
what computer it's initializing.

Replaced old constants with calls to dirky for the path.

This is pretty much the minimal required to get rid of those and get emacs to
start up successfully (on 'home' domain). Still need to hunt for any hiding in
work-domain-specific files.