Major Taskspace rework: Multiple taskspace groups.
Rewrote a lot of taskspace so that it can support, for instance, a 'work' and
'home' taskspace group with their separate settings and file/folder locations.

There is now a ':default' group, that has all settings defined in a defcustom.

Each group can override defaults, or not define anything to let that group use
the default.

Groups are stored in `taskspace/groups' - an alist of group symbol (:home),
display name ("Home Taskspace"), and settings.

Tried to switch to using f-el as much as I could.
  - Though it has some minor annoyances of its own.

Added dash as requirement so I could do list manip easier.

All taskspace functions are now group-aware, except the one or two that are so
basic or ignorant of context they don't need to be (e.g. org keyword search
one(s)).

Reworked use-package for taskspace in configure-project to have :home and :work
groups.